expand_env_var()をexpand_env_var.c内に移動
