+++ test/Makefile
@@ -16,7 +16,8 @@ SRCS = ../env.c \
 	   ../cmd_pid.c \
 	   ../cmd_cmd_invocation.c \
 	   ../cmd_cmd_invocation2.c \
-	   ../convert_ast2cmdinvo.c
+	   ../convert_ast2cmdinvo.c \
+	   ../expand_env_var.c
 HEADERS = ../env.h \
 		  ../parse.h \
 		  ../execution.h \
